add features nos novos arquivos na falha do merge
