+++ models/patient.py
@@ -1,5 +1,8 @@
 import pickle
 import pandas as pd
+import numpy as np
+import lime.lime_tabular
+import streamlit as st
 
 class Patient:
     """Classe de Paciente, responsável por armazenar os dados do paciente e realizar o dianóstico do mesmo, com a utilização de um model de ML.
@@ -53,6 +56,45 @@ class Patient:
                 columns=["Porcentagem"] # Necessário em português para visualização no front
             )
             return self.outputs[classification], prob_df
+
+    def explainer(self):
+        """Utiliza o LIME para explicação do predição do dianóstico
+
+        Returns:
+            pandas.Dataframe object: Dataframe contendo o valor de importância para cada atributo, seguinto o padrão [value] com cada atributo como index
+        """
+
+        path_database = st.secrets["path_database"]
+        database = pd.read_csv(path_database, sep=';', usecols=self.labels)
+
+        explainer = lime.lime_tabular.LimeTabularExplainer(
+            database.to_numpy(),
+            feature_names=self.labels,
+            class_names=self.outputs.keys(),
+            categorical_features=[count for count, value in enumerate(self.categorical_labels)],
+            categorical_names=self.categorical_labels,
+            kernel_width=3,
+            verbose=False
+        )
+
+        exp = explainer.explain_instance(
+            np.array(self.getRecord()[0]),
+            self.model.predict_proba,
+            num_features=14,
+            top_labels=3
+        )
+
+        for count, value in enumerate(self.saidas.keys()):
+            if value == self.classificacao:
+                pos_label = count
+
+        exp_dict = dict(exp.as_list(label=pos_label))
+
+        return pd.DataFrame(
+            exp_dict.values(),
+            columns=["Value"],
+            index=exp_dict.keys()
+        )
 
     # Setters
